fix: Cell 디자인 변경 사항 적용 (List와 통합)
PI-67189
