--- Sources/Montage/Components/List/List.swift
@@ -1,10 +1,0 @@
-//
-//  List.swift
-//  Montage
-//
-//  Created by Sanghoon Ahn on 8/14/24.
-//
-
-import SwiftUI
-
-public enum List {}
