+++ Sources/Montage/Components/Menu/Menu.swift
@@ -17,8 +17,8 @@ public struct Menu: View {
     public typealias MenuData = (
         title: String,
         isOn: Bool,
-        leftContent: Montage.List.Element.LeftContent?,
-        rightContent: Montage.List.Element.RightContent?
+        leftContent: Montage.Cell.LeftContent?,
+        rightContent: Montage.Cell.RightContent?
     )
     public typealias MenuDataSource = [(title: String?, datas: [MenuModel])]
 
@@ -26,26 +26,20 @@ public struct Menu: View {
     private let variant: Variant
     private let scroll: Bool
     private let backgroundColor: SwiftUI.Color
-    private let cellConfiguration: Montage.List.Cell.Configration
-    private let elementConfiguration: Montage.List.Element.Configuration
-    private let handler: ((Int, Int) -> Void)?
+    private let cell: (_ data: MenuModel, _ section: Int, _ index: Int) -> Cell
 
     public init(
         dataSources: MenuDataSource,
         variant: Variant = .normal,
         scroll: Bool = false,
         backgroundColor: SwiftUI.Color = .init(uiColor: .systemBackground),
-        cellConfiguration: Montage.List.Cell.Configration,
-        elementConfiguration: Montage.List.Element.Configuration,
-        handler: ((Int, Int) -> Void)? = nil
+        cell: @escaping (_ data: MenuModel, _ section: Int, _ index: Int) -> Cell
     ) {
         self.dataSources = dataSources
         self.variant = variant
         self.scroll = scroll
         self.backgroundColor = backgroundColor
-        self.cellConfiguration = cellConfiguration
-        self.elementConfiguration = elementConfiguration
-        self.handler = handler
+        self.cell = cell
     }
 
     public var body: some View {
@@ -69,16 +63,8 @@ public struct Menu: View {
                 Section {
                     ForEach(dataSource.datas.indices, id: \.self) { index in
                         let data = dataSource.datas[index]
-                        Cell(
-                            cellConfiguration: cellConfiguration,
-                            elementConfiguration: elementConfiguration,
-                            variant: variant,
-                            data: data,
-                            leftContent: data.leftContent,
-                            rightContent: data.rightContent
-                        ) {
-                            handler?(section, index)
-                        }
+                        cell(data, section, index)
+                            .padding(.horizontal, 12)
                     }
                 } header: {
                     if let title = dataSource.title {
@@ -88,41 +74,6 @@ public struct Menu: View {
                     }
                 }
             }
-        }
-    }
-    
-    private struct Cell: View {
-        var cellConfiguration: Montage.List.Cell.Configration
-        var elementConfiguration: Montage.List.Element.Configuration
-
-        var variant: Variant
-        var data: MenuModel
-        
-        var leftContent: List.Element.LeftContent?
-        var leftContentFromVariant: List.Element.LeftContent? {
-            switch variant {
-            case .normal: nil
-            case .radio: .radio(data.isOn)
-            case .checkbox: .check(data.isOn)
-            }
-        }
-        var rightContent: Montage.List.Element.RightContent?
-        var handler: (() -> Void)?
-        
-        var body: some View {
-            List.Cell(
-                model: .init(
-                    configuration: cellConfiguration,
-                    elementModel: .init(
-                        configuration: elementConfiguration,
-                        title: data.title,
-                        leftContent: leftContent == nil ? leftContentFromVariant : leftContent,
-                        rightContent: rightContent,
-                        handler: { handler?() }
-                    )
-                )
-            )
-            .padding(.horizontal, 12)
         }
     }
     
@@ -148,14 +99,14 @@ extension Montage.Menu {
     public struct MenuModel {
         public let title: String
         public var isOn: Bool
-        public var leftContent: Montage.List.Element.LeftContent?
-        public var rightContent: Montage.List.Element.RightContent?
+        public var leftContent: Montage.Cell.LeftContent?
+        public var rightContent: Montage.Cell.RightContent?
         
         public init(
             title: String,
             isOn: Bool,
-            leftContent: Montage.List.Element.LeftContent? = nil,
-            rightContent: Montage.List.Element.RightContent? = nil
+            leftContent: Montage.Cell.LeftContent? = nil,
+            rightContent: Montage.Cell.RightContent? = nil
         ) {
             self.title = title
             self.isOn = isOn
@@ -167,16 +118,16 @@ extension Montage.Menu {
 
 struct Menu_Previews: PreviewProvider {
     static var previews: some View {
-        let data = [
-            ("제목", ["1", "2", "3"]),
-            ("제목", ["1", "2", "3"]),
-            ("제목", ["1", "2", "3"]),
+        let data: [(String, [String])] = [
+            ("제목", ["Test1", "Test2", "Test3"]),
+            ("제목", ["Test1", "Test2", "Test3"]),
+            ("제목", ["Test1", "Test2", "Test3"]),
         ]
         let menudata: Menu.MenuDataSource = data.map { section in
             (
-                section.0,
-                section.1.map { row in
-                        .init(title: row, isOn: false, leftContent: nil, rightContent: nil)
+                title: section.0,
+                datas: section.1.map { row in
+                    Montage.Menu.MenuModel(title: row, isOn: false, leftContent: nil, rightContent: nil)
                 }
             )
         }
@@ -184,19 +135,15 @@ struct Menu_Previews: PreviewProvider {
             dataSources: menudata,
             variant: .normal,
             scroll: true,
-            cellConfiguration: .init(
-                padding: .normal,
-                paddingInset: false,
-                divider: false
-            ),
-            elementConfiguration: .init(
-                variant: .normal,
-                caption: nil,
-                bold: false,
-                disable: false
-            ),
-            handler: {
-                print("section: \($0), index: \($1)")
+            cell: { data, section, index in
+                Cell(title: data.title) {
+                    print("section: \(section), index: \(index)")
+                }
+                .padding(.normal)
+                .paddingInset(false)
+                .divider(false)
+                .bold(false)
+                .disable(false)
             }
         )
     }
